Fix the 'Insert do...end' snippet
This commit adds an additional placeholder in the `do...end` snippet.
Currently, if you want to create a `do...end` without any argument, you end up with the following:

``` ruby
my_method do ||
end
```

which makes the snippet unusable (writing RSpec things for instance).
With this commit, the whole ` |variable|` part is a placeholder. Then, when hitting `tab`, you move to the `variable` part and finally to the code part. It makes it more usable.